fix:コンポーネントをk理分けなど修正（Reviewed）
